requirements.txt
python-telegram-bot>=21.0,<22.0
python-dotenv>=1.0.1,<2.0.0
